changes to detection range
Detection range is now represented as an "abstract" class, it is a standalone script now. TowerRange and EnemyRange extends from detection range and should be used in their respective classes. Tower types are changed to extend area2d to give a hitbox. This sets the stage for ranged enemies and two-way interactions between towers and enemies