unexported some variables
godot does not support exporting of custom types, so a change of logic had to be made. for unit coupled to a detection range, their references are now obtained by calling getparent and getnode. This means that all detection range subtypes must be attatched to its owner scene by default.